Refactor group repository implementations in nested folders
