+++ client/src/ports/config/nightWatchConfigAdapters.js
@@ -1,4 +1,4 @@
-import NightWatchConfigEntity from '../domain/NightWatchConfigEntity';
+import NightWatchConfigEntity from '../../domain/NightWatchConfigEntity';
 
 export const toNightWatchConfigEntity = maxTotalTimeSpent =>
   new NightWatchConfigEntity({ maxTotalTimeSpent });
