+++ client/src/store/reducers/addCharacter.js
@@ -1,5 +1,5 @@
 import AddCharacterToParty from '../../usecases/party/AddCharacterToParty';
-import PartyRepository from '../../ports/PartyRepository';
+import PartyRepository from '../../ports/party/PartyRepository';
 
 const addCharacter = (state, { name }) => {
   const partyRepository = new PartyRepository(state);
